+++ common/mqtt_tests/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "mqtt_tests"
-version = "0.1.0"
+version = "0.4.0"
 edition = "2018"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
